GraphQL/requirements.txt
uvicorn==0.32.1
strawberry-graphql==0.254.0
starlette==0.41.3
requests==2.32.3
httpx==0.28.1